엔티티별 Repository 작성 (EmailAuthRepository 추가)
