+++ src/main/java/com/example/gatherplan/appointment/repository/EmailAuthRepository.java
@@ -1,0 +1,14 @@
+package com.example.gatherplan.appointment.repository;
+
+import com.example.gatherplan.appointment.repository.entity.EmailAuth;
+
+import java.util.Optional;
+
+public interface EmailAuthRepository {
+
+    void save(EmailAuth emailAuth);
+
+    void delete(String email);
+
+    Optional<EmailAuth> findEmailAuthByEmail(String email);
+}
